requirements.txt
pyyaml
dotenv
